Transition to IPJ, Iroha rc2 version update (#330)
* Transition to IPJ

* Iroha rc1 update

* Test fixes

* Iroha rc2 update

* Updated permissions, roles, testing account

* Test fixes

* Update batches

* Fixed status streaming and batch testing

* Switched to ordered batches and to Query API of IPJ

* More queries refactoring, changed wrapper

+++ notary-commons/src/main/kotlin/notary/IrohaCommand.kt
@@ -48,7 +48,7 @@ sealed class IrohaCommand {
     data class CommandCreateAsset(
         val assetName: String,
         val domainId: String,
-        val precision: Short
+        val precision: Int
     ) : IrohaCommand()
 
     /**
